Refactor Deb functionality into packaging package
